backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=4.5.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
django-allauth>=0.57.0
dj-rest-auth>=4.0.0
drf-spectacular>=0.26.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
transformers>=4.30.0 